+++ sberbook/core/subscription/src/main/java/ru/sberbook/sberbookroot/SubscriptionController.java
@@ -4,6 +4,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Flux;
+import ru.sberbook.sberbookroot.entities.SubscribtionsEntity;
 
 import java.util.List;
