+++ sberbook/core/subscription/src/main/java/ru/sberbook/sberbookroot/SubscriptionService.java
@@ -1,6 +1,7 @@
 package ru.sberbook.sberbookroot;
 
 import reactor.core.publisher.Flux;
+import ru.sberbook.sberbookroot.entities.SubscribtionsEntity;
 
 import java.util.List;
