+++ sberbook/core/subscription/src/main/java/ru/sberbook/sberbookroot/SubscriptionServiceImpl.java
@@ -2,8 +2,11 @@ package ru.sberbook.sberbookroot;
 
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
+import ru.sberbook.sberbookroot.entities.SubscribersEntity;
+import ru.sberbook.sberbookroot.entities.SubscribtionsEntity;
+import ru.sberbook.sberbookroot.repositories.SubscriberRepository;
+import ru.sberbook.sberbookroot.repositories.SubscriptionRepository;
 
-import java.util.Arrays;
 import java.util.List;
 
 import static java.util.stream.LongStream.range;
